Ising fitter for hompolymers in jupyter
Fits a capped hompolymeric set of NRC-type repeats.  This notebook contains data conversion, fit equation generation, and fitting, along with plotting, bootstrap analysis, and bootstrap statistics (including correlation plots).  

Runs in python 2.7.  It does not run in python 3 due to a change in eval()

Run with the data in the NRC_data folder